refactor: Add press/api/version_upgrade.py - Move api endpoints and logic linked to site version upgrade to the new file from press/api/site.py
